refactor(repository): Reduce memory allocations during manifest compaction (#4084)
* Change struct for tracking committed content

Committed content only ever has a value of 'true' for committed so use
an empty struct and an existance check instead.

* Don't copy committed manifest set for compaction

Assuming the number of committed manifests is much larger than the
number of manifest updates, it seems reasonable to update the logic to
write manifests out to not delete entries from the set being operated
on. Doing so allows us to avoid creating a duplicate of the set of all
committed manifests during compaction, which could also save some memory
as the temporary map wasn't being right-sized based on the the number of
committed entries. This also works because writing data either fails or
succeeds completely. That means there's no possibility of only some
entries being written out but not others, which means callers can use
the presence of an error to determine how to react (i.e. clear pending
set).

* Benchmarks for compaction